pyproject.toml
[project]
name = "sebba-code"
version = "0.1.0"
description = "LangGraph coding agent with git-native memory"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.4",
    "langchain-core>=0.3",
    "langchain-anthropic>=0.3",
    "langchain-openai>=0.3",
    "langchain>=0.3",
    "pyyaml>=6.0",
    "click>=8.0",
]

[project.scripts]
sebba-code = "sebba_code.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sebba_code"]

[dependency-groups]
dev = [
    "pytest>=8.0",
]
